refaktorer vilkårsvurderingdetails og splitt opp varslinger/meldinger
