+++ src/components/vilkårsvurdering-details/VilkårsvurderingDetails.tsx
@@ -22,14 +22,16 @@ import Personalia from '../../types/Personalia';
 import BrukerFyller18ÅrIPeriodenMelding from '../bruker-fyller-18-år-i-perioden-melding/BrukerFyller18ÅrIPeriodenMelding';
 import BrukerUnder18ÅrIHelePeriodenMelding from '../bruker-under-18-år-i-hele-perioden-melding/BrukerUnder18ÅrIHelePeriodenMelding';
 import dayjs from 'dayjs';
+import SøknadsVarslinger from '../søknads-varslinger/SøknadsVarslinger';
+import VilkårsVarslinger from '../vilkårs-varslinger/VilkårsVarslinger';
 
 interface VilkårsvurderingDetailsProps {
-    søknadResponse: Behandling;
+    behandling: Behandling;
     personalia: Personalia;
 }
 
-const VilkårsvurderingDetails = ({
-    søknadResponse: {
+const VilkårsvurderingDetails = ({ behandling, personalia: { fødselsdato } }: VilkårsvurderingDetailsProps) => {
+    const {
         tiltakspengerYtelser,
         statligeYtelser,
         kommunaleYtelser,
@@ -37,66 +39,17 @@ const VilkårsvurderingDetails = ({
         lønnsinntekt,
         institusjonsopphold,
         barnetillegg,
-        registrerteTiltak,
         alderVilkårsvurdering,
-        søknad: { startdato, sluttdato, fritekst },
-    },
-    personalia: { fødselsdato },
-}: VilkårsvurderingDetailsProps) => {
-    const dagpengePerioder = statligeYtelser.finnDagpengeperioder();
-    const aapPerioder = statligeYtelser.finnAapPerioder();
-    const kvpPerioder = kommunaleYtelser.finnKvpPerioderFraSøknaden();
-    const introPerioder = kommunaleYtelser.finnIntroprogramPerioderFraSøknaden();
-    const pensjonsordningPerioder = pensjonsordninger.finnPensjonsordningerOppgittISøknaden();
-    const lønnsinntektPerioder = lønnsinntekt.finnLønnsinntektOppgittISøknaden();
-    const institusjonsoppholdPerioder = institusjonsopphold.finnInstitusjonsoppholdOppgittISøknaden();
-    const tiltakspengerPerioder = tiltakspengerYtelser.finnPerioderTilManuellVurdering();
-
-    const periodeBrukerIkkeHarFylt18År = alderVilkårsvurdering.finnPeriodeHvorBrukerIkkeHarFylt18År();
-    const datoBrukerFyller18År = dayjs(fødselsdato).add(18, 'years');
-    const brukerErUnder18ÅrIHelePerioden = datoBrukerFyller18År.isAfter(dayjs(sluttdato));
-
-    const visManglendeRegistrertTiltakMelding = !registrerteTiltak || registrerteTiltak.length === 0;
-    const visMeldingOmAtBrukerFyller18ÅrIPerioden = !brukerErUnder18ÅrIHelePerioden && periodeBrukerIkkeHarFylt18År;
+        søknad: { fritekst },
+    } = behandling;
 
     return (
         <div className={styles.vilkårsvurderingDetails}>
             <Heading level="1" size="small">
                 Søknad
             </Heading>
-            <Alert variant="info" fullWidth style={{ marginTop: '1rem' }}>
-                Foreløpig har vi ikke alle opplysninger til å vurdere søknaden.
-            </Alert>
-            {fritekst && (
-                <Alert variant="info" fullWidth style={{ marginTop: '1rem' }}>
-                    Bruker har fylt ut tilleggsopplysninger i søknaden.
-                </Alert>
-            )}
-            {brukerErUnder18ÅrIHelePerioden && <BrukerUnder18ÅrIHelePeriodenMelding />}
-            {visMeldingOmAtBrukerFyller18ÅrIPerioden && (
-                <BrukerFyller18ÅrIPeriodenMelding periode={periodeBrukerIkkeHarFylt18År} />
-            )}
-            {visManglendeRegistrertTiltakMelding && (
-                <Alert variant="warning" fullWidth style={{ marginTop: '1rem', paddingBottom: 0 }}>
-                    <strong>
-                        Det er ikke registrert tiltak på bruker i perioden {formatDate(startdato)} -{' '}
-                        {formatDate(sluttdato)}
-                    </strong>
-                    <p>Søknaden trenger manuell behandling</p>
-                </Alert>
-            )}
-            {tiltakspengerPerioder.length > 0 && (
-                <HarStatligYtelseMelding perioder={tiltakspengerPerioder} ytelseText="Tiltakspenger" />
-            )}
-            {dagpengePerioder.length > 0 && (
-                <HarStatligYtelseMelding perioder={dagpengePerioder} ytelseText="Dagpenger" />
-            )}
-            {aapPerioder.length > 0 && <HarStatligYtelseMelding perioder={aapPerioder} ytelseText="AAP" />}
-            {kvpPerioder.length > 0 && <HarKvpDeltakelseMelding />}
-            {introPerioder.length > 0 && <HarIntroprogramDeltakelseMelding perioder={introPerioder} />}
-            {pensjonsordningPerioder.length > 0 && <HarPensjonsordningMelding perioder={pensjonsordningPerioder} />}
-            {institusjonsoppholdPerioder.length > 0 && <HarInstitusjonsoppholdMelding />}
-            {lønnsinntektPerioder.length > 0 && <HarLønnsinntektMelding perioder={lønnsinntektPerioder} />}
+            <SøknadsVarslinger fritekst={fritekst} />
+            <VilkårsVarslinger behandling={behandling} fødselsdato={fødselsdato} />
             <div style={{ marginTop: '4rem' }}>
                 <Accordion title="Alder (§3)">
                     <AlderVilkårsvurderingTable
